+++ mern/client/src/screens/RecoverSession/recoverSession.js
@@ -1,8 +1,7 @@
 import React, { Component } from "react";
 import "./recoverSession.css";
 import GenericButton from "../../Components/Buttons/GenericButton";
-import { withRouter } from "../../withRouter";
-
+import { withRouter } from "../withRouter";
 class RecoverSession extends Component {
   constructor(props) {
     super(props);
